Switch to iterator get commits & fix blocking run
I did this for performance reasons. First I switch to a Sequence based
iterator solution so that it only parses commits as needed on demand.

After adding the iterator we were still having a weird hanging issue so
we dug further in and found out that we had a weird blocking issue with
larger output. So we switched to a stream processing model for the
output and it resolved this blocking issue.

I also made the default max infinite instead of 300 as it isn't needed
anymore thanks to the permission fixes.

ps-id: 815BD163-71F5-4F9D-AC48-3F8B07A79EB2

[changelog]
fixed: issue blocking git commands with large output
changed: fetch commits to return Sequence w/ an Iterator, not an Array