Add summary to the Commit type
I did this so that we could use it to implement the planned `--commits`
switch to show the list of commit sha1s and summary for each of the
commits in either unreleased, or latest sub-commands.

Note: I am not leaving a changelog entry in this commit because it is an
intern only change. There is no user facing impact with this change.

ps-id: FB807990-9E31-44D7-BE23-CE81E5028FFD