Fix bug where first commit was being reprocessed
I did this because it was causing the end-user to see double the
changelog entries for the first commit in some cases. This is a crappy
experience and doesn't build trust so I decided to fix it.

This was happening because we were missing a guard statement at the top
of the next() method in the Iterator that checked if we had already
exhausted the content.

Beyond that I removed a conditional in one of the logic paths that is no
longer needed because of the new guard statement.

[changelog]
fixed: bug where double the first commit entries would show up