Implements SwiftArgumentParser for command line options
This implements additional subcommands that allow for quick and easy
parsing.

This implements a new action system and flow to handle subcommands with
additional parameters.

[changelog]
added: Swift Argument Parser to allow for a cleaner appraoch to command line arguments.
added: Implemented support for full, released <optional_version> and unreleased
added: Implemented a --commits and to released and unreleased to allow all of the commit info to be returned
removed: Removed all of the sub commands options that were present
changed: Refactored some code and cleaned things up

ps-id: 871C7F84-1801-4681-9EED-5396CA29F7D2