Make `released` sub-command require release-id
We did this because we weren't sure of what the use case was for getting
all of the released changelog entries or commits. Also we couldn't get
Apple's CLI parser lib to properly allow the release id argument to be
optional. It would allow us to set it and report it as optional in the
help. However, when you would run it without that argument it would
error out saying it was missing a required argument. So, since we didn't
think the use case without the release id was valuable we just
eliminated it.

[changelog]
removed: variant of released with no release-id argument

ps-id: 91612DCD-CFC9-4916-8ECF-E5B9AC7F19B4