Fix up the diff links
This effectively rewrites the `full` sub-command to make it work with
the new ChangelogCommits sequence and take into account the link
referencing in the markdown file so that users can quickly get the diffs
of the changes they are looking for in github.

[changelog]
fixed: Diff links now display all releases and the headers of each section correctly links to the diff.

ps-id: 65A9B70B-BE53-42B8-B35C-621E238124C4