Change releases to be identified by tags
I did this because we found there were issues with releases being
identified by information in the commit message body. For example if a
release: v1.2.3 was included in a commit message body and then that
commit was cherry picked to another branch. Should that other branch
have the release v1.2.3. The answer is know and it breaks things. So, we
thought about it more and decided the correct thing to do was switch to
tags matching the pattern of ^v?\d+\.\d+\.\d+$ identifying a releases.
Tags that don't match that pattern are just ignored.

[changelog]
removed: 'release' commit message body entries as supported entry types
changed: to identify releases by Git tags matching pattern: ^v?\d+\.\d+\.\d+$

ps-id: F6FED912-DE05-48AB-89A4-AB5E88DFFF01