Fix crash when in repo without a repo url
I did this because when you would use the full sub-command in a repo
that didn't have a base repo url set it would crash. That is a horrible
experience. So instead now it safely degrades back to a mode without the
link references at the bottom of the Changelog output.

[changelog]
fixed: crash when repo doesn't have a repo url

ps-id: CFB4BAF9-04D1-490F-A1A4-00D7653DFFD3