Rework how release fetching is implemented
Previously release fetching worked by starting at HEAD and walking down
the git tree from there. Identifying releases as it went down the tree.
This turned out to be problematic because it could miss out on hotfix
releases that are in separate branches. This also had the problem of
potentially getting incorrect changelog entries for versions that were
merged into the path that was being walked.

So we restructured it so that it fetches all the release tags and then
gets the details for each of those releases and composes the overall
changelog. This makes it so that the changelog includes all of the
release across the entire git tree. It also makes it so that the
changelog entries associated with each one are correct.

[changelog]
fixed: bug where hotfix releases would be missed
fixed: bug where merged release branches could cause incorrect changelog entries