Replace hard coded "master" refs with cur branch
I did this so that the commands that previously depended on the hard
coded "master" branch will now run against the currently checked out
branch. This enables these commands to work on repositories that use
"main" or other branches for their mainline as well as enables these
commands to be used on other branches.

[changelog]
fixed: issue where commands were hard coded for "master" to use the checked out branch